% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA24204159.R
\name{calculate_ES_cpp}
\alias{calculate_ES_cpp}
\title{Calculate ES using R}
\usage{
calculate_ES_cpp(distribution, p, n = 10000, ...)
}
\arguments{
\item{distribution}{a character string specifying the distribution (for example, "norm" corresponds to the standard normal)}

\item{p}{a real valued parameter taking values in (0, 1)}

\item{n}{the number of parts dividing the interval}

\item{...}{other parameters}
}
\value{
the value of TVaR
}
\description{
Calculate ES for common distribution using R
}
\examples{
\dontrun{
    ES_normal <- calculate_ES_cpp("norm", 0.95,mean=100,sd=5)
    print(ES_normal)
    ES_exp <- calculate_ES_cpp("exp", 0.95,rate=5)
    print(ES_exp)
}
}
