% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA24204159.R
\name{calculate_TVaR_cpp}
\alias{calculate_TVaR_cpp}
\title{Calculate TVaR using R}
\usage{
calculate_TVaR_cpp(distribution, p, n = 10000, ...)
}
\arguments{
\item{distribution}{a character string specifying the distribution (for example, "norm" corresponds to the standard normal)}

\item{p}{a real valued parameter taking values in (0, 1)}

\item{n}{the number of parts dividing the interval}

\item{...}{other parameters}
}
\value{
the value of TVaR
}
\description{
Calculate TVaR for common distribution using R
}
\examples{
\dontrun{
    TVaR_normal <- calculate_TVaR_cpp("norm", 0.95,mean=100,sd=5)
    print(TVaR_normal)
    TVaR_exp <- calculate_TVaR_cpp("exp", 0.95,rate=5)
    print(TVaR_exp)
}
}
