% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_integral_sum}
\alias{compute_integral_sum}
\title{the inner product of two vectors using cpp}
\usage{
compute_integral_sum(x_vals, y_vals)
}
\arguments{
\item{x_vals}{one of the vectors}

\item{y_vals}{the other one of the vectors}
}
\value{
the value of the inner product
}
\description{
add every components of two vectors and sum
}
\examples{
\dontrun{
compute_integral_sum(c(1,2),c(1,2))
}
}
