\name{SA24204159-package}
\alias{SA24204159-package}
\alias{SA24204159}
\docType{package}
\title{
  Functions for calculating risk measures for common distributions
}
\description{
  Several R and Rcpp function are included in this package for calculating risk measures for common distributions.
}
\author{
WenHan Qian

Maintainer: WenHan Qian <qwh_0424@mail.ustc.edu.cn>
}

\keyword{ package }

